Renamed 'Types' module to 'DataTypes' for clarity
